Add minimum timestep monitoring and stopping criterion
Adds dt_min tracking and a new MinDtAccidents stopping criterion class to
detect and handle situations where the adaptive timestep becomes so small that
it leads to performance issues.

The dt_min violations are recorded alongside the stopping criteria, with
idx=-1. Indeing for requested stopping criteria is updated accordingly to start
with -2.